python/requirements.txt
annorepo-client
elasticsearch
jsonpickle
loguru
pyyaml
requests
tqdm
